+++ reclutch/src/display/gpu/mod.rs
@@ -1,56 +1,27 @@
+mod cmd;
+mod pipe;
+
 use {
     super::*,
+    cmd::*,
     image::ConvertBuffer,
     lyon::tessellation::{self as tess, basic_shapes as shapes},
+    pipe::*,
     raw_window_handle::HasRawWindowHandle,
     std::{collections::HashMap, error::Error},
 };
 
-const SAMPLE_COUNT: u32 = 8;
-
-struct CommandGroupData {
-    display: Vec<DisplayCommand>,
-    render: Vec<RenderCommand>,
-
-    bounds: Rect,
-
-    // TODO(jazzfool): combine these
-    vbo: wgpu::Buffer,
-    ibo: wgpu::Buffer,
-    ubo: wgpu::Buffer,
-
-    locals_bind_group: wgpu::BindGroup,
+pub(crate) mod common {
+    pub use {
+        super::{cmd::*, pipe::*, *},
+        crate::error,
+        lyon::tessellation as tess,
+    };
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct Mesh {
-    base_vertex: u32,
-    indices: std::ops::Range<u32>,
-}
+pub(crate) const SAMPLE_COUNT: u32 = 8;
 
-#[derive(Debug)]
-enum RenderCommand {
-    DrawMesh {
-        mesh: Mesh,
-        paint: RenderPaint,
-    },
-    DrawBackdropFilter {
-        clip: Mesh,
-        filter_pipe: wgpu::RenderPipeline,
-    },
-    DrawMeshToStencil {
-        mesh: Mesh,
-    },
-    Save,
-    SaveTexture(f32),
-    Restore,
-    Translate,
-    Scale,
-    Rotate,
-    Clear,
-}
-
-fn copy_to_array<T: Sized + Default + Copy>(src: &[T]) -> [T; 32] {
+pub(crate) fn copy_to_array<T: Sized + Default + Copy>(src: &[T]) -> [T; 32] {
     assert!(
         src.len() <= 32,
         "WGPU implementation only supports up to 32 gradient stops; received {}",
@@ -62,148 +33,24 @@ fn copy_to_array<T: Sized + Default + Copy>(src: &[T]) -> [T; 32] {
     array
 }
 
-fn color_to_bgra(color: Color) -> [f32; 4] {
+pub(crate) fn color_to_bgra(color: Color) -> [f32; 4] {
     [color.blue, color.green, color.red, color.alpha]
 }
 
-fn color_to_rgba(color: Color) -> [f32; 4] {
+pub(crate) fn color_to_rgba(color: Color) -> [f32; 4] {
     [color.red, color.green, color.blue, color.alpha]
-}
-
-#[derive(Debug, Default, Clone, Copy, PartialEq)]
-#[repr(C)]
-struct GpuGradientStop {
-    pos: f32,
-    _pad: [u32; 3],
-    color: [f32; 4],
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-#[repr(C)]
-struct GpuGradient {
-    start: [f32; 2],
-    end: [f32; 2],
-    stops: [GpuGradientStop; 32],
-    length: u32,
-}
-
-impl GpuGradient {
-    fn new(gradient: &Gradient) -> Self {
-        let mut stops: Vec<_> = gradient
-            .stops
-            .iter()
-            .map(|(pos, color)| GpuGradientStop {
-                pos: *pos,
-                _pad: [0; 3],
-                color: color_to_rgba(*color),
-            })
-            .collect();
-
-        let length = stops.len() as u32;
-        stops.resize(
-            32,
-            GpuGradientStop {
-                pos: 0.0,
-                _pad: [0; 3],
-                color: [0.0; 4],
-            },
-        );
-
-        let stops = copy_to_array(&stops);
-
-        GpuGradient {
-            start: gradient.start.to_array(),
-            end: gradient.end.to_array(),
-            stops,
-            length,
-        }
-    }
-}
-
-#[derive(Debug)]
-enum RenderPaint {
-    Fill,
-    LinearGradient(wgpu::BindGroup),
-    RadialGradient(wgpu::BindGroup),
-    Image(wgpu::BindGroup),
-}
-
-impl RenderPaint {
-    fn from_style_color(
-        sc: StyleColor,
-        device: &wgpu::Device,
-        pipelines: &PipelineData,
-    ) -> (Self, Option<wgpu::Buffer>) {
-        match sc {
-            StyleColor::Color(_) => (RenderPaint::Fill, None),
-            StyleColor::LinearGradient(ref gradient) => {
-                let gradient = GpuGradient::new(gradient);
-                let buf = device
-                    .create_buffer_mapped(1, wgpu::BufferUsage::UNIFORM)
-                    .fill_from_slice(&[gradient]);
-                let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    layout: &pipelines.gradient_bind_group_layout,
-                    bindings: &[wgpu::Binding {
-                        binding: 0,
-                        resource: wgpu::BindingResource::Buffer {
-                            buffer: &buf,
-                            range: 0..std::mem::size_of::<GpuGradient>() as _,
-                        },
-                    }],
-                });
-
-                (RenderPaint::LinearGradient(bind_group), buf.into())
-            }
-            StyleColor::RadialGradient(ref gradient) => {
-                let gradient = GpuGradient::new(gradient);
-                let buf = device
-                    .create_buffer_mapped(1, wgpu::BufferUsage::UNIFORM)
-                    .fill_from_slice(&[gradient]);
-                let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    layout: &pipelines.gradient_bind_group_layout,
-                    bindings: &[wgpu::Binding {
-                        binding: 0,
-                        resource: wgpu::BindingResource::Buffer {
-                            buffer: &buf,
-                            range: 0..std::mem::size_of::<GpuGradient>() as _,
-                        },
-                    }],
-                });
-                (RenderPaint::RadialGradient(bind_group), buf.into())
-            }
-        }
-    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Default)]
 #[repr(C)]
 #[repr(align(16))]
-struct Vertex {
+pub(crate) struct Vertex {
     pos: [f32; 2],
     tex_coord: [f32; 2],
     color: [f32; 4],
 }
 
-#[derive(Debug, Clone, Copy)]
-#[repr(C)]
-struct Uniforms {
-    ortho: nalgebra::Matrix4<f32>,
-    transform: nalgebra::Matrix4<f32>,
-}
-
-#[derive(Debug, Clone, Copy)]
-#[repr(C)]
-struct Locals {
-    transform: nalgebra::Matrix4<f32>,
-}
-
-#[derive(Debug)]
-enum Resource {
-    Image(wgpu::TextureView),
-    Font(wgpu_glyph::GlyphBrush<'static, ()>),
-}
-
-struct VertexCtor(Color, Option<Rect>);
+pub(crate) struct VertexCtor(Color, Option<Rect>);
 
 impl tess::VertexConstructor<tess::FillVertex, Vertex> for VertexCtor {
     fn new_vertex(&mut self, vertex: tess::FillVertex) -> Vertex {
@@ -239,248 +86,6 @@ impl tess::VertexConstructor<tess::StrokeVertex, Vertex> for VertexCtor {
     }
 }
 
-fn clip_to_gd_item(clip: &DisplayClip) -> GraphicsDisplayItem {
-    unimplemented!()
-}
-
-fn convert_stroke(stroke: GraphicsDisplayStroke) -> tess::StrokeOptions {
-    tess::StrokeOptions::default()
-        .with_line_width(stroke.thickness)
-        .with_line_cap(match stroke.cap {
-            LineCap::Flat => tess::LineCap::Butt,
-            LineCap::Square => tess::LineCap::Square,
-            LineCap::Round => tess::LineCap::Round,
-        })
-        .with_line_join(match stroke.join {
-            LineJoin::Miter => tess::LineJoin::Miter,
-            LineJoin::Round => tess::LineJoin::Round,
-            LineJoin::Bevel => tess::LineJoin::Bevel,
-        })
-        .with_miter_limit(stroke.miter_limit)
-}
-
-fn tessellate(gd_item: &GraphicsDisplayItem) -> (Vec<Vertex>, Vec<u16>) {
-    let mut buffer: tess::VertexBuffers<Vertex, u16> = Default::default();
-
-    match gd_item {
-        GraphicsDisplayItem::Line { a, b, stroke } => shapes::stroke_polyline(
-            [*a, *b].iter().copied(),
-            false,
-            &convert_stroke(stroke.clone()),
-            &mut tess::BuffersBuilder::new(
-                &mut buffer,
-                VertexCtor(stroke.color.color_or_black(), None),
-            ),
-        ),
-        GraphicsDisplayItem::Rectangle { rect, paint } => match paint {
-            GraphicsDisplayPaint::Fill(color) => shapes::fill_rectangle(
-                rect,
-                &Default::default(),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(color.color_or_black(), None),
-                ),
-            ),
-            GraphicsDisplayPaint::Stroke(stroke) => shapes::stroke_rectangle(
-                rect,
-                &convert_stroke(stroke.clone()),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(stroke.color.color_or_black(), None),
-                ),
-            ),
-        },
-        GraphicsDisplayItem::RoundRectangle { rect, radii, paint } => match paint {
-            GraphicsDisplayPaint::Fill(color) => shapes::fill_rounded_rectangle(
-                rect,
-                &shapes::BorderRadii::new(radii[0], radii[1], radii[2], radii[3]),
-                &Default::default(),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(color.color_or_black(), None),
-                ),
-            ),
-            GraphicsDisplayPaint::Stroke(stroke) => shapes::stroke_rounded_rectangle(
-                rect,
-                &shapes::BorderRadii::new(radii[0], radii[1], radii[2], radii[3]),
-                &convert_stroke(stroke.clone()),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(stroke.color.color_or_black(), None),
-                ),
-            ),
-        },
-        GraphicsDisplayItem::Ellipse {
-            center,
-            radii,
-            paint,
-        } => match paint {
-            GraphicsDisplayPaint::Fill(color) => shapes::fill_ellipse(
-                *center,
-                *radii,
-                Angle::zero(),
-                &Default::default(),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(color.color_or_black(), None),
-                ),
-            ),
-            GraphicsDisplayPaint::Stroke(stroke) => shapes::stroke_ellipse(
-                *center,
-                *radii,
-                Angle::zero(),
-                &convert_stroke(stroke.clone()),
-                &mut tess::BuffersBuilder::new(
-                    &mut buffer,
-                    VertexCtor(stroke.color.color_or_black(), None),
-                ),
-            ),
-        },
-        GraphicsDisplayItem::Image { dst, .. } => shapes::fill_rectangle(
-            dst,
-            &Default::default(),
-            &mut tess::BuffersBuilder::new(
-                &mut buffer,
-                VertexCtor(Color::new(0.0, 0.0, 0.0, 1.0), Some(*dst)),
-            ),
-        ),
-    };
-
-    (buffer.vertices, buffer.indices)
-}
-
-fn load_shader(
-    shader: &str,
-    filename: &str,
-    kind: shaderc::ShaderKind,
-    compiler: &mut shaderc::Compiler,
-    format: wgpu::TextureFormat,
-    device: &wgpu::Device,
-) -> Result<wgpu::ShaderModule, error::GpuError> {
-    let options = shaderc::CompileOptions::new()
-        .ok_or_else(|| error::GpuError::CompilerError("failed to create compile options".into()))?;
-    let spirv = compiler
-        .compile_into_spirv(shader, kind, filename, "main", Some(&options))
-        .map_err(|err| error::GpuError::CompilerError(err.description().into()))?;
-    Ok(device.create_shader_module(spirv.as_binary()))
-}
-
-fn create_pipeline(
-    vs: &wgpu::ShaderModule,
-    fs: &wgpu::ShaderModule,
-    bind_group_layouts: &[&wgpu::BindGroupLayout],
-    device: &wgpu::Device,
-    format: wgpu::TextureFormat,
-) -> wgpu::RenderPipeline {
-    let layout =
-        device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor { bind_group_layouts });
-    device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-        layout: &layout,
-        vertex_stage: wgpu::ProgrammableStageDescriptor {
-            module: vs,
-            entry_point: "main",
-        },
-        fragment_stage: Some(wgpu::ProgrammableStageDescriptor {
-            module: fs,
-            entry_point: "main",
-        }),
-        rasterization_state: Some(wgpu::RasterizationStateDescriptor {
-            front_face: wgpu::FrontFace::Ccw,
-            cull_mode: wgpu::CullMode::None,
-            depth_bias: 0,
-            depth_bias_slope_scale: 0.0,
-            depth_bias_clamp: 0.0,
-        }),
-        primitive_topology: wgpu::PrimitiveTopology::TriangleList,
-        color_states: &[wgpu::ColorStateDescriptor {
-            format,
-            color_blend: wgpu::BlendDescriptor::REPLACE,
-            alpha_blend: wgpu::BlendDescriptor::REPLACE,
-            write_mask: wgpu::ColorWrite::ALL,
-        }],
-        depth_stencil_state: None,
-        index_format: wgpu::IndexFormat::Uint16,
-        vertex_buffers: &[wgpu::VertexBufferDescriptor {
-            step_mode: wgpu::InputStepMode::Vertex,
-            stride: std::mem::size_of::<Vertex>() as _,
-            attributes: &[
-                wgpu::VertexAttributeDescriptor {
-                    format: wgpu::VertexFormat::Float2,
-                    offset: 0,
-                    shader_location: 0,
-                },
-                wgpu::VertexAttributeDescriptor {
-                    format: wgpu::VertexFormat::Float2,
-                    offset: 8,
-                    shader_location: 1,
-                },
-                wgpu::VertexAttributeDescriptor {
-                    format: wgpu::VertexFormat::Float4,
-                    offset: 16,
-                    shader_location: 2,
-                },
-            ],
-        }],
-        sample_count: SAMPLE_COUNT,
-        sample_mask: !0,
-        alpha_to_coverage_enabled: true,
-    })
-}
-
-fn create_render_paint(
-    gd_item: &GraphicsDisplayItem,
-    device: &wgpu::Device,
-    pipelines: &PipelineData,
-    resources: &HashMap<u64, Resource>,
-) -> Result<(RenderPaint, Option<wgpu::Buffer>), error::DisplayError> {
-    match gd_item {
-        GraphicsDisplayItem::Line { stroke, .. } => Ok(RenderPaint::from_style_color(
-            stroke.color.clone(),
-            device,
-            pipelines,
-        )),
-        GraphicsDisplayItem::Rectangle { paint, .. }
-        | GraphicsDisplayItem::RoundRectangle { paint, .. }
-        | GraphicsDisplayItem::Ellipse { paint, .. } => match paint {
-            GraphicsDisplayPaint::Fill(color) => Ok(RenderPaint::from_style_color(
-                color.clone(),
-                device,
-                pipelines,
-            )),
-            GraphicsDisplayPaint::Stroke(stroke) => Ok(RenderPaint::from_style_color(
-                stroke.color.clone(),
-                device,
-                pipelines,
-            )),
-        },
-        GraphicsDisplayItem::Image { resource, .. } => {
-            if let ResourceReference::Image(ref id) = resource {
-                if let Some(Resource::Image(ref texture_view)) = resources.get(id) {
-                    let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-                        layout: &pipelines.image_bind_group_layout,
-                        bindings: &[
-                            wgpu::Binding {
-                                binding: 0,
-                                resource: wgpu::BindingResource::TextureView(texture_view),
-                            },
-                            wgpu::Binding {
-                                binding: 1,
-                                resource: wgpu::BindingResource::Sampler(&pipelines.image_sampler),
-                            },
-                        ],
-                    });
-
-                    Ok((RenderPaint::Image(bind_group), None))
-                } else {
-                    Err(error::DisplayError::InvalidResource(*id))
-                }
-            } else {
-                Err(error::DisplayError::MismatchedResource(resource.id()))
-            }
-        }
-    }
-}
-
 fn create_msaa_framebuffer(
     device: &wgpu::Device,
     swap_chain_desc: &wgpu::SwapChainDescriptor,
@@ -507,134 +112,10 @@ fn create_msaa_framebuffer(
         .create_default_view()
 }
 
-fn upload_to_buffer<T: 'static + Sized + Copy>(
-    device: &wgpu::Device,
-    queue: &mut wgpu::Queue,
-    contents: &[T],
-    buffer: &wgpu::Buffer,
-) {
-    let temp_buf = device
-        .create_buffer_mapped(contents.len(), wgpu::BufferUsage::COPY_SRC)
-        .fill_from_slice(contents);
-    let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor { todo: 0 });
-    encoder.copy_buffer_to_buffer(
-        &temp_buf,
-        0,
-        buffer,
-        0,
-        (std::mem::size_of::<T>() * contents.len()) as _,
-    );
-    queue.submit(&[encoder.finish()]);
-}
-
-#[derive(Debug)]
-struct PipelineData {
-    fill_pipeline: wgpu::RenderPipeline,
-    linear_gradient_pipeline: wgpu::RenderPipeline,
-    radial_gradient_pipeline: wgpu::RenderPipeline,
-    image_pipeline: wgpu::RenderPipeline,
-
-    globals_bind_group: wgpu::BindGroup,
-    locals_bind_group_layout: wgpu::BindGroupLayout,
-    gradient_bind_group_layout: wgpu::BindGroupLayout,
-    image_bind_group_layout: wgpu::BindGroupLayout,
-    image_sampler: wgpu::Sampler,
-}
-
-#[derive(Debug)]
-enum SaveState<T: std::fmt::Debug, L: std::fmt::Debug> {
-    Transform(T),
-    Layer(L),
-}
-
-#[derive(Debug)]
-struct SaveStack<T: std::fmt::Debug, L: std::fmt::Debug> {
-    stack: Vec<SaveState<T, L>>,
-    top_transform: Option<usize>,
-    top_layer: Option<usize>,
-}
-
-impl<T: std::fmt::Debug, L: std::fmt::Debug> Default for SaveStack<T, L> {
-    fn default() -> Self {
-        SaveStack {
-            stack: Vec::new(),
-            top_transform: None,
-            top_layer: None,
-        }
-    }
-}
-
-impl<T: std::fmt::Debug, L: std::fmt::Debug> SaveStack<T, L> {
-    fn new() -> Self {
-        Default::default()
-    }
-
-    fn save(&mut self, state: SaveState<T, L>) {
-        *match &state {
-            SaveState::Transform(_) => &mut self.top_transform,
-            SaveState::Layer(_) => &mut self.top_layer,
-        } = Some(self.stack.len());
-
-        self.stack.push(state);
-    }
-
-    fn restore(&mut self) -> Option<SaveState<T, L>> {
-        let state = if self.stack.len() > 0 {
-            Some(self.stack.remove(self.stack.len() - 1))
-        } else {
-            None
-        };
-
-        match &state {
-            Some(SaveState::Transform(_)) => {
-                self.top_transform = self.stack.iter().rev().position(|state| {
-                    if let SaveState::Transform(_) = state {
-                        true
-                    } else {
-                        false
-                    }
-                })
-            }
-            Some(SaveState::Layer(_)) => {
-                self.top_layer = self.stack.iter().rev().position(|state| {
-                    if let SaveState::Layer(_) = state {
-                        true
-                    } else {
-                        false
-                    }
-                })
-            }
-            _ => (),
-        };
-
-        state
-    }
-
-    fn peek_transform(&self) -> Option<&T> {
-        self.top_transform.map(|idx| {
-            if let SaveState::Transform(transform) = self.stack.get(idx).unwrap() {
-                transform
-            } else {
-                panic!()
-            }
-        })
-    }
-
-    fn peek_layer(&self) -> Option<&L> {
-        self.top_layer.map(|idx| {
-            if let SaveState::Layer(layer) = self.stack.get(idx).unwrap() {
-                layer
-            } else {
-                panic!()
-            }
-        })
-    }
-}
-
 pub struct GpuGraphicsDisplay {
     command_groups: linked_hash_map::LinkedHashMap<u64, CommandGroupData>,
     next_command_group_id: u64,
-    resources: HashMap<u64, Resource>,
+    resources: ResourceMap,
     next_resource_id: u64,
 
     globals: Uniforms,
@@ -971,8 +452,8 @@ impl GpuGraphicsDisplay {
 
                 layout(location = 0) out vec4 o_color;
 
-                layout(set = 1, binding = 0) uniform texture2D u_texture;
-                layout(set = 1, binding = 1) uniform sampler u_sampler;
+                layout(set = 2, binding = 0) uniform texture2D u_texture;
+                layout(set = 2, binding = 1) uniform sampler u_sampler;
 
                 void main() {
                     o_color = texture(sampler2D(u_texture, u_sampler), f_tex_coord);
@@ -1115,7 +596,7 @@ impl GraphicsDisplay for GpuGraphicsDisplay {
                     dimension: wgpu::TextureDimension::D2,
                     format: wgpu::TextureFormat::Bgra8Unorm,
                     sample_count: 1,
-                    usage: wgpu::TextureUsage::SAMPLED,
+                    usage: wgpu::TextureUsage::SAMPLED | wgpu::TextureUsage::COPY_DST,
                     mip_level_count: 1,
                     array_layer_count: 1,
                 });
@@ -1138,13 +619,13 @@ impl GraphicsDisplay for GpuGraphicsDisplay {
                     },
                     wgpu::TextureCopyView {
                         texture: &texture,
-                        mip_level: 1,
+                        mip_level: 0,
+                        array_layer: 0,
                         origin: wgpu::Origin3d {
                             x: 0.0,
                             y: 0.0,
                             z: 0.0,
                         },
-                        array_layer: 1,
                     },
                     wgpu::Extent3d {
                         width,
@@ -1185,101 +666,16 @@ impl GraphicsDisplay for GpuGraphicsDisplay {
     ) -> Result<CommandGroupHandle, Box<dyn std::error::Error>> {
         let id = self.next_command_group_id;
 
-        let mut vertices = Vec::new();
-        let mut indices = Vec::new();
-        let mut render = Vec::new();
-
-        let locals = Locals {
-            transform: nalgebra::Matrix4::identity(),
-        };
-
-        let mut save_count = 0;
-
-        for cmd in commands {
-            match cmd {
-                DisplayCommand::Item(item) => match item {
-                    DisplayItem::Graphics(gd_item) => {
-                        let paint = create_render_paint(
-                            gd_item,
-                            &self.device,
-                            &self.pipelines,
-                            &self.resources,
-                        )
-                        .map_err(|err| Box::new(err))?
-                        .0;
-
-                        let (mesh_vertices, mesh_indices) = tessellate(gd_item);
-
-                        let mesh = Mesh {
-                            base_vertex: vertices.len() as _,
-                            indices: indices.len() as _..(indices.len() + mesh_indices.len()) as _,
-                        };
-
-                        vertices.extend(mesh_vertices.into_iter());
-                        indices.extend(mesh_indices.into_iter());
-
-                        render.push(RenderCommand::DrawMesh { mesh, paint });
-                    }
-                    DisplayItem::Text(td_item) => {}
-                },
-                DisplayCommand::BackdropFilter(clip, filter) => {}
-                DisplayCommand::Save => {
-                    save_count += 1;
-                    render.push(RenderCommand::Save);
-                }
-                DisplayCommand::SaveLayer(opacity) => {
-                    save_count += 1;
-                    render.push(RenderCommand::SaveTexture(*opacity));
-                }
-                DisplayCommand::Restore => {
-                    if save_count > 0 {
-                        save_count -= 1;
-                        render.push(RenderCommand::Restore);
-                    } // else error?
-                }
-                _ => (),
-            }
-        }
-
-        if protected.unwrap_or(true) {
-            for i in 0..save_count {
-                render.push(RenderCommand::Restore);
-            }
-        }
-
-        let ubo = self
-            .device
-            .create_buffer_mapped(1, wgpu::BufferUsage::UNIFORM)
-            .fill_from_slice(&[locals]);
-        let locals_bind_group = self.device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &self.pipelines.locals_bind_group_layout,
-            bindings: &[wgpu::Binding {
-                binding: 0,
-                resource: wgpu::BindingResource::Buffer {
-                    buffer: &ubo,
-                    range: 0..std::mem::size_of::<Locals>() as _,
-                },
-            }],
-        });
-
         self.next_command_group_id += 1;
         self.command_groups.insert(
             id,
-            CommandGroupData {
-                display: commands.to_vec(),
-                render,
-                bounds: display_list_bounds(commands).map_err(|err| Box::new(err))?,
-                vbo: self
-                    .device
-                    .create_buffer_mapped(vertices.len(), wgpu::BufferUsage::VERTEX)
-                    .fill_from_slice(&vertices),
-                ibo: self
-                    .device
-                    .create_buffer_mapped(indices.len(), wgpu::BufferUsage::INDEX)
-                    .fill_from_slice(&indices),
-                ubo,
-                locals_bind_group,
-            },
+            CommandGroupData::new(
+                commands,
+                &self.device,
+                &self.pipelines,
+                &self.resources,
+                protected,
+            )?,
         );
 
         Ok(CommandGroupHandle::new(id))
